requirements.txt
alabaster==0.7.12
arrow==1.2.2
asgiref==3.5.0
Babel==2.10.1
binaryornot==0.4.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cookiecutter==1.7.3
crispy-bootstrap5==0.6
cryptography==37.0.2
Deprecated==1.2.13
dj-static==0.0.6
Django==4.0.2
django-browser-reload==1.3.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-filter==21.1
django-oauth-plus==2.2.9
django-oauth-toolkit==2.0.0
django-oauth2-provider==0.2.6.1
django-simple-history==3.0.0
django-sslserver==0.22
django-tailwind==3.1.1
djangorestframework==3.13.1
djangorestframework-oauth==1.1.0
docutils==0.17.1
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.3
Jinja2==3.1.2
jinja2-time==0.2.0
jwcrypto==1.2
Markdown==3.3.6
MarkupSafe==2.1.1
numpy==1.22.3
oauth2==1.9.0.post1
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
poyo==0.5.0
psycopg2-binary==2.9.3
Pycco==0.6.0
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.8
pystache==0.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-slugify==6.1.2
pytz==2022.1
requests==2.27.1
shortuuid==1.0.8
six==1.16.0
smartypants==2.0.1
snowballstemmer==2.2.0
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
static3==0.7.0
text-unidecode==1.3
tzdata==2021.5
urllib3==1.26.9
wrapt==1.14.1
zipp==3.7.0
